moved AnnotationFinder from core.impl to core package
