+++ salt-api/src/main/java/org/corpus_tools/salt/core/impl/SNodeImpl.java
@@ -24,6 +24,7 @@ import java.util.Iterator;
 import java.util.List;
 import java.util.Set;
 
+import org.corpus_tools.salt.core.AnnotationFinder;
 import org.corpus_tools.salt.core.SAbstractAnnotation;
 import org.corpus_tools.salt.core.SAnnotation;
 import org.corpus_tools.salt.core.SFeature;
